+++ client/remotepower-agent-win.py
@@ -48,7 +48,7 @@ import time
 import urllib.request
 import urllib.error
 
-VERSION = '6.2.0'
+VERSION = '6.2.1'
 DEFAULT_POLL = 60